+++ src/generic-table/interfaces/gt-information.ts
@@ -1,8 +1,8 @@
 export interface GtInformation {
-  rowLength:number
-  currentPage:number
-  pagesTotal:number
-  filter:boolean | Object
-  refresh:any
-  searchTerms?:string
+  rowLength: number;
+  currentPage: number;
+  pagesTotal: number;
+  filter: boolean | Object;
+  refresh: any;
+  searchTerms?: string;
 }