+++ src/generic-table/interfaces/gt-row.ts
@@ -1,0 +1,3 @@
+export interface GtRow {
+  isOpen?: boolean;
+}