+++ src/generic-table/pipes/get-property.pipe.ts
@@ -1,13 +1,14 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import {GtConfigField} from "../interfaces/gt-config-field";
 import {GtConfigSetting} from '../interfaces/gt-config-setting';
+import { GtRow } from '../interfaces/gt-row';
 
 @Pipe({
   name: 'getProperty'
 })
-export class GetPropertyPipe implements PipeTransform {
+export class GetPropertyPipe<R extends GtRow> implements PipeTransform {
 
-  transform(config: Array<GtConfigField> | Array<GtConfigSetting>, objectKey: string, property: string, refresh?:boolean): any {
+  transform(config: GtConfigField<R>[] | GtConfigSetting[], objectKey: string, property: string, refresh?: boolean): any {
     try {
       var output = null;
       for (var i = 0; i < config.length;i++){