+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Release History
 ---------------
 ## [Unreleased]
+
+# [1.0.0-rc.7] - 2016-11-22
 ### Breaking changes
 - Moved search (true/false) to `GtConfigSetting` i.e. to disable global search for a field use `GtConfigSetting` and to defined a search function use `GtConfigField`.
 
@@ -9,6 +11,10 @@ Release History
 
 ### Fixed
 - Column content (i.e td content) is now wrapped within a span, click etc. is now triggered by the span and not the td element.
+
+# [1.0.0-rc.6] - 2016-11-22
+### Breaking changes
+- Added strong typing for all interfaces (you can always use `<any>` where type is required)
 
 # [1.0.0-rc.5] - 2016-11-18
 ### Breaking changes